+++ client/src/components/idea/discussion/comment-parser.js
@@ -1,7 +1,9 @@
+import styled from "@emotion/styled";
 import React from "react";
 import replace from "react-string-replace";
 import tinycolor from "tinycolor2";
-import {UiBadge, UiPrettyUsername} from "ui";
+import {UiBadge, UiClickableTip, UiKeyboardInput, UiPrettyUsername} from "ui";
+import {UiButton} from "ui/button";
 
 const parseComment = (message, moderatorsData, tagsData) => {
     const regex = /[^{}]+(?=})/g;
@@ -29,6 +31,8 @@ const parseTag = (result, moderatorsData, tagsData, index) => {
         return parseBoardTagData(data, tagsData, index);
     } else if (data[0] === "data_user") {
         return parseModeratorData(data, index);
+    } else if (data[0] === "data_diff_view") {
+        return parseDiffView(data, index);
     }
 };
 
@@ -45,6 +49,24 @@ const parseModeratorData = (data, index) => {
     return <span className={"font-weight-bold"} key={data[1] + index}>
         <UiPrettyUsername user={{id: parseInt(data[1]), username: data[2]}}/>
     </span>
+};
+
+const DiffViewButton = styled.div`
+  display: inline-block;
+  color: var(--font-color);
+  cursor: pointer;
+`;
+
+const parseDiffView = (data, index) => {
+    return <DiffViewButton>
+        <UiClickableTip id={data[1] + index} title={"Diff View"} description={<React.Fragment>
+            <strong>From:</strong>
+            <UiKeyboardInput>{data[2]}</UiKeyboardInput>
+            <br/>
+            <strong>To:</strong>
+            <UiKeyboardInput>{data[3]}</UiKeyboardInput>
+        </React.Fragment>} icon={<UiButton tiny={true} label={"View Diff"}>{data[1]}</UiButton>}/>
+    </DiffViewButton>
 };
 
 export default parseComment;