+++ server/src/main/java/net/feedbacky/app/data/board/webhook/WebhookExecutor.java
@@ -68,6 +68,7 @@ public class WebhookExecutor {
       case IDEA_OPEN:
       case IDEA_CLOSE:
       case IDEA_EDIT:
+        embedBuilder.addField("Title", data.get(WebhookMapData.IDEA_TITLE.getName()), true);
         embedBuilder.addField("Description", data.get(WebhookMapData.IDEA_DESCRIPTION.getName()), true);
         break;
       case IDEA_COMMENT:
@@ -116,7 +117,7 @@ public class WebhookExecutor {
   }
 
   public enum WebhookMapData {
-    USER_NAME("user_name"), USER_AVATAR("user_avatar"), USER_ID("user_id"), IDEA_NAME("idea_name"),
+    USER_NAME("user_name"), USER_AVATAR("user_avatar"), USER_ID("user_id"), IDEA_NAME("idea_name"), IDEA_TITLE("idea_title"),
     IDEA_DESCRIPTION("idea_description"), IDEA_LINK("idea_link"), IDEA_ID("idea_id"), COMMENT_DESCRIPTION("comment_description"), COMMENT_ID("comment_id"),
     TAGS_CHANGED("tags_changed"), CHANGELOG_NAME("changelog_name"), CHANGELOG_DESCRIPTION("changelog_description");